flowsint-api/pyproject.toml
[tool.poetry]
name = "flowsint-api"
version = "0.1.0"
license = "AGPL-3.0-or-later"
description = "API server for flowsint"
authors = ["dextmorgn <contact@flowsint.io>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
flowsint-core = { path = "../flowsint-core", develop = true }
flowsint-types = { path = "../flowsint-types", develop = true }
flowsint-transforms = { path = "../flowsint-transforms", develop = true }
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
redis = "^5.0"
celery = "^5.3"
python-dotenv = "^1.0"
python-jose = {extras = ["cryptography"], version = "^3.4"}
requests = "^2.31"
pydantic = "^2.0"
neo4j = "^5.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
asyncpg = "^0.30"
alembic = "1.13.0"
passlib = {extras = ["bcrypt"], version = "^1.7"}
bcrypt = ">=4.0.0,<5.0.0"
sse-starlette = "^1.8"
networkx = "^2.6.3"
email-validator = "^2.2.0"
mistralai = "^1.9.3"
python-multipart = "^0.0.20"
openpyxl = "^3.1.2"
jsonschema = "^4.25.1"
openai = "^1.54.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
isort = "^5.12"
flake8 = "^6.0"
mypy = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
